requirements.txt
keras
pillow
numpy
tensorflow
matplotlib
ollama